pyproject.toml
[project]
name = "llm-review-sys-sft"
version = "0.1.0"
description = "Paper Review Model SFT Training based on Qwen2.5-7B-Instruct"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llamafactory",
    "transformers>=4.45.0",
    "datasets>=2.16.0",
    "accelerate>=0.30.0",
    "peft>=0.11.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.43.0",
    "scipy",
    "einops",
    "sentencepiece",
    "tiktoken",
    "numpy",
    "modelscope>=1.35.1",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "openai>=1.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
]
eval = [
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "openai>=1.0.0",
    "tqdm>=4.65.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
